+++ src/main/java/com/github/jlangch/venice/impl/specialforms/SpecialForms_OtherFunctions.java
@@ -37,6 +37,7 @@ import com.github.jlangch.venice.VncException;
 import com.github.jlangch.venice.impl.Destructuring;
 import com.github.jlangch.venice.impl.IFormEvaluator;
 import com.github.jlangch.venice.impl.ISequenceValuesEvaluator;
+import com.github.jlangch.venice.impl.IVeniceInterpreter;
 import com.github.jlangch.venice.impl.InterruptChecker;
 import com.github.jlangch.venice.impl.Modules;
 import com.github.jlangch.venice.impl.debug.agent.DebugAgent;
@@ -646,7 +647,8 @@ public class SpecialForms_OtherFunctions {
                     .doc(
                         "Runs the expr count times in the most effective way and returns a list of " +
                         "elapsed nanoseconds for each invocation. It's main purpose is supporting " +
-                        "benchmark test.")
+                        "benchmark test.\n\n" +
+                        "*Note:* All macros in the expression are expanded upfront.")
                     .examples(
                          "(dobench 1000 (+ 1 1))",
                          "(dobench 1000 2 1000 (+ 1 1))")
@@ -693,11 +695,17 @@ public class SpecialForms_OtherFunctions {
                     final VncVal expr = args.fourth();
                     final VncFunction statusFn = Coerce.toVncFunction(args.nth(4));
 
+                    final IVeniceInterpreter interpreter = ctx.getInterpreter();
+                    final VncVal exprMacroExpanded = interpreter.macroexpand_all(
+                                                        new CallFrame("dobench", args, specialFormMeta),
+                                                        expr,
+                                                        env);
+
                     return Benchmark.benchmark(
                                 warmUpIterations,
                                 gcRuns,
                                 iterations,
-                                expr,
+                                exprMacroExpanded,
                                 statusFn,
                                 env,
                                 ctx.getEvaluator());
@@ -718,9 +726,9 @@ public class SpecialForms_OtherFunctions {
                         "supporting benchmark tests. Returns the expression result of the last " +
                         "invocation.\n\n" +
                         "*Note:*¶" +
-                        "The expression is evaluated for every run. " +
-                        "Alternatively a zero or one arg function referenced by a symbol can be " +
-                        "passed:\n\n" +
+                        "All macros in the expression are expanded upfront. The expression is evaluated " +
+                        "for every run. Alternatively a zero or one arg function referenced by a symbol " +
+                        "can be passed:\n\n" +
                         "```                      \n" +
                         "(let [f (fn [] (+ 1 1))] \n" +
                         "  (dorun 10 f))          \n" +
@@ -775,10 +783,16 @@ public class SpecialForms_OtherFunctions {
                 }
 
                 try {
-                    final VncVal first = evaluator.evaluate(expr, env, false);
+                    final IVeniceInterpreter interpreter = ctx.getInterpreter();
+                    final VncVal exprMacroExpanded = interpreter.macroexpand_all(
+                                                        new CallFrame("dorun", args, specialFormMeta),
+                                                        expr,
+                                                        env);
+
+                    final VncVal first = evaluator.evaluate(exprMacroExpanded, env, false);
 
                     for(int ii=1; ii<count; ii++) {
-                        final VncVal result = evaluator.evaluate(expr, env, false);
+                        final VncVal result = evaluator.evaluate(exprMacroExpanded, env, false);
 
                         InterruptChecker.checkInterrupted(Thread.currentThread(), "dorun");